services/governance-api/pyproject.toml
[project]
name = "governance-api"
version = "0.2.0"
description = "AI Governance Worley UAIP UC3 — AI Governance & Incident Resolution Hub Incident Resolution Hub"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.30.0,<1.0",
    "httpx>=0.27.0,<1.0",
    "pydantic>=2.0,<3.0",
    "azure-identity>=1.17.0,<2.0",
    "azure-monitor-query>=1.3.0,<2.0",
    "azure-servicebus>=7.13.0,<8.0",
    "azure-cosmos>=4.7.0,<5.0",
    "opentelemetry-api>=1.25.0,<2.0",
    "opentelemetry-sdk>=1.25.0,<2.0",
    "opentelemetry-exporter-otlp>=1.25.0,<2.0",
    "azure-monitor-opentelemetry-exporter>=1.0.0b25,<2.0",
    "pyjwt[crypto]>=2.8.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.24.0,<1.0",
    "httpx>=0.27.0,<1.0",
    "respx>=0.21.0,<1.0",
    "ruff>=0.4.0,<1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py311"
